Pull chat header out of chat app.
